// src/features/face-auth/types/face-auth.ts
export type FaceQualityReason =
  | 'NO_FACE'
  | 'MULTIPLE_FACES'
  | 'FACE_TOO_SMALL'
  | 'FACE_TOO_LARGE'
  | 'FACE_MOVED'
  | 'LOW_FRAME_RATE'
  | 'LOW_SIGNAL'
  | 'TOO_FEW_FRAMES'
  | 'TOO_MANY_FRAMES'